pyproject.toml
[tool.poetry]
name = "ruuvitag-collector"
version = "1.0.0"
description = "Collect measurements from RuuviTag sensors to InfluxDB and other databases."
authors = ["Niko Korhonen <1004138+niktheblak@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
ruuvitag-sensor = "^0.13.0"
influxdb = "^5.2.3"
google-cloud-pubsub = "^1.1.0"
google-cloud-logging = "^1.14.0"
retrying = "^1.3.3"
confuse = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
